requirements.txt
argon2-cffi==16.3.0
cffi==1.10.0
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mongoengine==0.13.0
passlib==1.7.1
pycparser==2.18
PyJWT==1.5.3
pymongo==3.5.1
python-magic==0.4.13
six==1.10.0
Werkzeug==0.12.2
